pyproject.toml
# ==============================================================================
# The [project] table defined by Official python.org
#
# Read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
# ==============================================================================
[project]
name = "pythonic_gmail"
# Increment version before each release - follow `semantic versioning <https://semver.org/>`_
# Currently, poetry 2.1.X doesn't support dynamic versioning
# (Read https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#version)
# So this value has to be aligned with the one in ``pythonic_gmail/_version.py``
version = "0.1.2"
description = "An example project generated by cookiecutter-pywf_open_source."
# Read https://dev-exp-share.readthedocs.io/en/latest/search.html?q=Pick+An+Open+Source+License+For+Python+Project&check_keywords=yes&area=default
# To pick a license and update the ``license``, ``classifier`` field in ``pyproject.toml``
# And also update the ``LICENSE.txt`` file in the git repo.
license = "MIT"
license-files = ["LICENSE.txt", "AUTHORS.rst"]
authors = [
    { name = "Sanhe Hu", email = "husanhe@email.com" },
]
maintainers = [
    { name = "Sanhe Hu", email = "husanhe@email.com" },
]
keywords = []
readme = "README.rst"
requires-python = ">=3.10,<4.0"
# Full list of classifiers: https://pypi.org/classifiers/
classifier = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: Unix",
    # Note: classfier like "Programming Language :: Python :: 3 :: Only",
    # "Programming Language :: Python :: 3.X" is no longer needed,
    # because we have ``requires-python`` field.
]

# ------------------------------------------------------------------------------
# Core Dependencies
# ------------------------------------------------------------------------------
dependencies = [
    "typing-extensions>=4.0.0,<5.0.0; python_version<'3.11'",
    "more-itertools>=10.6.0; python_version<'3.12'",
    "soft-deps>=0.1.1,<1.0.0",
    "enum-mate>=0.1.1,<1.0.0",
    "func-args>=1.0.1,<2.0.0",
    "iterproxy>=0.3.1,<1.0.0",
    "google-api-python-client>=2.157.0,<3.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "google-auth-oauthlib>=1.2.1,<2.0.0",
]

# ------------------------------------------------------------------------------
# Optional dependency that can be used in ``pip install ${your_project_name}[${feature_name}]``
# Sometime this is also called "extras"
# Read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies
#
# IMPORTANT: all optional dependencies has to be compatible with the "requires-python" field
# ------------------------------------------------------------------------------
[project.optional-dependencies]
email = [
    "beautifulsoup4>=4.10.0,<5.0.0", # parse HTML email
]

# ------------------------------------------------------------------------------
# Local Development dependenceies
# ------------------------------------------------------------------------------
dev = [
    "twine>=6.0.0,<7.0.0", # distribute package to PyPI
    "wheel>=0.45.0,<1.0.0", # build wheel package
    "build>=1.2.1,<2.0.0", # build source distribution
    "rich>=13.8.1,<14.0.0", # pretty print
    "google-api-python-client-stubs>=1.29.0,<2.0.0",
]

# ------------------------------------------------------------------------------
# (Unit/Coverage/Integration/Load) Test dependenceies
# ------------------------------------------------------------------------------
test = [
    "pytest>=8.2.2,<9.0.0", # Testing framework
    "pytest-cov>=6.0.0,<7.0.0", # Coverage reporting
]

# ------------------------------------------------------------------------------
# Documentation build dependenceies
# ------------------------------------------------------------------------------
doc = [
    "Sphinx>=7.4.7,<8.0.0",
    "sphinx-copybutton>=0.5.2,<1.0.0", # add copy button to code block
    "sphinx-design>=0.6.1,<1.0.0", # add additional design pattern to sphinx
    "sphinx-jinja>=2.0.2,<3.0.0", # enable jinja syntax in reStructuredText
    "furo==2024.8.6", # the furo sphinx theme
    "pygments>=2.18.0,<3.0.0", # syntax highlight
    "ipython>=8.18.1,<8.19.0", # interactive Python
    "nbsphinx>=0.8.12,<1.0.0", # add jupyter notebook in sphinx doc
    "rstobj==1.2.1", # generate reStructuredText from Python code
    "docfly==3.0.0", # automaticall generate .. toctree directives and API reference doc
]

# ------------------------------------------------------------------------------
# Automation (devops) dependenceies
# ------------------------------------------------------------------------------
auto = [
]

# Quick Links
[project.urls]
Homepage = "https://github.com/MacHu-GWU/pythonic_gmail-project"
Documentation = "https://pythonic-gmail.readthedocs.io/en/latest/"
Repository = "https://github.com/MacHu-GWU/pythonic_gmail-project"
Issues = "https://github.com/MacHu-GWU/pythonic_gmail-project/issues"
Changelog = "https://github.com/MacHu-GWU/pythonic_gmail-project/blob/main/release-history.rst"
Download = "https://pypi.org/pypi/pythonic-gmail#files"

# For command line interface, read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts
[project.scripts]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.9.0,<2.0.0"

[tool.poetry]
# Files to include in the package distribution
packages = [
    { include = "pythonic_gmail", from = ".", to = "." }
]
# Files to exclude from the package
exclude = [
    "pythonic_gmail/tests/*",
    "pythonic_gmail/docs/*",
    "**/*.pyc",
    "**/*.pyo",
]

# python workflow tool config
[tool.pywf]
# The specific python version you use for local development
dev_python = "3.11.8"
# --- github.com
github_account = "MacHu-GWU"
# Create GitHub token in https://github.com/settings/tokens and put the token in
# ``${HOME}/home_secret.json``
github_token_field = "providers.github.accounts.sh.users.sh.secrets.dev.value"
# --- codecov.io (for code coverage test report)
codecov_account = "MacHu-GWU"
# Create Codecov token in https://app.codecov.io/account/gh/${codecov_account}/access
# and put the token in ``${HOME}/home_secret.json``
codecov_token_field = "providers.codecov_io.accounts.sh.users.sh.secrets.dev.value"
# --- readthedocs.org (for documentation hosting)
# Create Readthedocs token in https://app.readthedocs.org/accounts/tokens/
# and put the token at ``${HOME}/home_secret.json``
readthedocs_token_field = "providers.readthedocs.accounts.sh.users.sh.secrets.dev.value"
# Readthedocs project name, usually it is the same as your project name
readthedocs_project_name = "pythonic_gmail"

# Read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#declaring-the-build-backend
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
